Add `wat2wasm` example compatible with Embedded Swift
