+++ Sources/WasmParser/CMakeLists.txt
@@ -1,14 +1,6 @@
 add_wasmkit_library(WasmParser
-  Stream/ByteStream.swift
-  Stream/FileHandleStream.swift
-  Stream/Stream.swift
-  BinaryInstructionDecoder.swift
-  InstructionVisitor.swift
-  LEB.swift
-  ParsingLimits.swift
-  WasmParser.swift
-  WasmTypes.swift
+  FileHandleStream.swift
 )
 
 target_link_wasmkit_libraries(WasmParser PUBLIC
-  WasmTypes SystemPackage)
+  WasmParserCore)
