+++ Sources/WasmParserCore/BinaryInstructionDecoder.swift
