+++ Sources/WasmParserCore/ParsingLimits.swift
