+++ Sources/WasmParserCore/Stream/ByteStream.swift
