+++ Sources/WasmParserCore/Stream/Stream.swift
